+++ R/rdf_to_rwtbl.R
@@ -3,15 +3,18 @@
 #' 
 #' `rdf_to_rwtbl()` converts an rdf list to a tibble.
 #' 
-#' The rdf list is converted to a data frame, and then converted to a 
-#' [tibble::tibble()]. All of the `meta` entries into the rdf list
+#' The rdf object is converted to a data frame, and then converted to a 
+#' [tibble::tibble()]. All of the `meta` entries in the rdf object
 #' are stored as attributes in the returned tibble. These attributes are:
 #' `mrm_config_name`, `owner`, `description`, `create_date`, and `n_traces`.
 #' 
-#' @param rdf An rdf object (from [read.rdf()]).
-#' @param scenario An optional parameter, that if it is not `NULL` (default)
-#'   will be added to the tibble as another variable. Typically a string, but it 
-#'   is not coerced to a string.
+#' If the rdf contains a scalar slot(s), the scalar slot value(s) will be 
+#' repeated for every timestep.
+#' 
+#' @param rdf An rdf object (from [read_rdf()]).
+#' @param scenario An optional parameter, that if it is not `NULL` or `NA` 
+#'   (default) will be added to the tibble as another variable. Coerced to a 
+#'   character if it is not already a character.
 #' @param keep_cols Either boolean, or a character vector of column names to 
 #'   keep in the returned tibble. The values of `keep_cols` work as follows:
 #'   * `FALSE` (default) only includes the defaults columns: `Timestep`, 
@@ -37,24 +40,29 @@
 #' rdftbl <- rdf_to_rwtbl(keyRdf, add_ym = FALSE, keep_cols = "Object")
 #' rdftbl <- rdf_to_rwtbl(sysRdf, scenario = "ISM1988_2014,2007Dems,IG,2002")
 #' 
-#' @seealso [read.rdf()]
+#' # rdf_to_rwtbl2 wants a file path instead of an rdf object
+#' rdfPath <- system.file(
+#'   "extdata/Scenario/ISM1988_2014,2007Dems,IG,Most/KeySlots.rdf", 
+#'   package = "RWDataPlyr"
+#' )
+#' rdftbl <- rdf_to_rwtbl2(rdfPath)
+#' 
+#' @seealso [read_rdf()]
 #' 
 #' @export
 
 rdf_to_rwtbl <- function(rdf, scenario = NULL, keep_cols = FALSE, add_ym = TRUE)
 {
-  stopifnot(methods::is(rdf, "rdf"))
-  
-  stopifnot(
-    (is.logical(keep_cols) && !is.na(keep_cols) && length(keep_cols == 1)) || 
-    (is.character(keep_cols) && length(keep_cols) > 0)
+  .Deprecated(
+    msg = paste0(
+      "`rdf_to_rwtbl()` is deprecated and will be removed in a future release.\n",
+      "Consider using `rdf_to_rwtbl2()` instead.\n",
+      "See help(\"rdf_to_rwtbl()\"), as it uses different arguments."
+    )
   )
+  stopifnot(is_rdf(rdf))
   
-  stopifnot(is.logical(add_ym) && !is.na(add_ym) && length(add_ym) == 1)
-  
-  # this function only reads in one scenario, so should only have one 
-  # scenario name
-  stopifnot(is.null(scenario) || length(scenario) == 1)
+  check_rdf_to_rwtbl_args(scenario, keep_cols, add_ym, "rdf_to_rwtbl")
   
   # rdf[["meta"]] contains meta data
   # rdf[["runs"]] will be the length of rdf[[1]]$number_of_runs
@@ -69,8 +77,11 @@ rdf_to_rwtbl <- function(rdf, scenario = NULL, keep_cols = FALSE, add_ym = TRUE)
   if ((is.logical(keep_cols) && !keep_cols) || is.character(keep_cols))
     tbl <- select_rdftbl_cols(tbl, keep_cols)
   
-  if (!is.null(scenario))
+  if (!is.null(scenario)) {
+    if (!is.character(scenario))
+      scenario <- as.character(scenario)
     tbl$Scenario <- scenario
+  }
   
   if (add_ym)
     tbl <- add_ym_to_rdftbl(tbl)
@@ -83,6 +94,71 @@ rdf_to_rwtbl <- function(rdf, scenario = NULL, keep_cols = FALSE, add_ym = TRUE)
     "create_date" = atts$create_date,
     "n_traces" = nRun
   )
+}
+
+#' `rdf_to_rwtbl2()` converts an rdf file into a tibble, but is 
+#' faster than `rdf_to_rwtbl()` since it uses c++. It also reads the rdf file
+#' in, while `rdf_to_rwtbl()` needs an `rdf` object.
+#' 
+#' @param file The relative or absolute rdf filename.
+#'  
+#' @rdname rdf_to_rwtbl
+#' 
+#' @export
+rdf_to_rwtbl2 <- function(file, scenario = NA_character_, keep_cols = FALSE, 
+                          add_ym = TRUE)
+{
+  if (! is.character(file) & length(file) != 1) {
+    stop(
+      "In `rdf_to_rwtbl2()`, `file` should be a single entry character vector,",
+      "  i.e., a file path to one rdf file.",
+      call. = FALSE
+    )
+  }
+  
+  rdf_vec <- read_rdf(file, rdf = FALSE)
+  
+  check_rdf_to_rwtbl_args(scenario, keep_cols, add_ym, "rdf_to_rwtbl2")
+
+  add_cols <- c("ObjectName", "SlotName", "ObjectType" ,"Unit", 
+                "RulesetFileName", "InputDMIName")
+
+  if (is.logical(keep_cols)) {
+    if (keep_cols) {
+      keep_cols <- c(req_rwtbl_cols(), add_cols)
+    } else {
+      keep_cols <- req_rwtbl_cols()
+    }
+  } else {
+    # combine keep_cols with the required columns
+    keep_cols <- keep_cols[!(keep_cols %in% req_rwtbl_cols())]
+    keep_cols <- c(req_rwtbl_cols(), keep_cols)
+  }
+  
+  # check that keep_cols are all expected column names
+  missing_cols <- keep_cols[!(keep_cols %in% c(req_rwtbl_cols(), add_cols))]
+  if (length(missing_cols) > 0)
+    warning(
+      "The following columns specified by 'keep_cols' were not found in the rwtbl:\n",
+      "    ", toString(missing_cols)
+    )
+  
+  if (is.null(scenario)) {
+    scenario <- NA_character_
+  }
+  
+  if (! is.character(scenario))
+    scenario <- as.character(scenario)
+  
+  t1 <- rdf_to_rwtbl_cpp(
+    rdf_vec, 
+    scenario = scenario, 
+    keep_cols = keep_cols,
+    add_ym = add_ym
+  ) %>% 
+    tibble::as_tibble()
+  
+  t1
 }
 
 
@@ -202,4 +278,51 @@ rwtbl_get_atts <- function(rwtbl)
     c("mrm_config_name", "owner", "description", "create_date", "n_traces"), 
     names(attributes(rwtbl))
   )]
+}
+
+
+check_rdf_to_rwtbl_args <- function(scenario, keep_cols, add_ym, call_func)
+{
+  if (is.logical(keep_cols)) {
+    if (length(keep_cols) != 1 || is.na(keep_cols)) {
+      stop(
+        "In `", call_func, "`(), if `keep_cols`, is logical, it should have ",
+        "a length of 1, and only be TRUE or FALSE.",
+        call. = FALSE
+      )
+    }
+  } else if (is.character(keep_cols)) {
+    if ( ! (length(keep_cols) > 0)) {
+      stop(
+        "In `", call_func, "`(), if `keep_cols`, is a character, it should ",
+        "have a length of 1 or more.",
+        call. = FALSE
+      )
+    }
+  } else {
+    stop(
+      "In `", call_func, "`(), `keep_cols` should either be a character or ",
+      "boolean.",
+      call. = FALSE
+    )
+  }
+  
+  if (length(add_ym) != 1 || !is.logical(add_ym) || is.na(add_ym)) {
+    stop(
+      "In `", call_func, "`(), `add_ym` should be a length 1 boolean.",
+      call. = FALSE
+    )
+  }
+  
+  # this function only reads in one scenario, so should only have one 
+  # scenario name
+  if (!is.null(scenario)) {
+    if (length(scenario) != 1) {
+      stop(
+        "In `", call_func, "`(), `scenario`, should either be NULL, or have a ",
+        "length of 1.",
+        call. = FALSE
+      )
+    }
+  }
 }