+++ R/read_rdf.R
@@ -146,11 +146,13 @@ read_rdf_run <- function(rdf.mat, rdf.obj)
 #' `read.rdf2()` is deprecated and will be removed in a future release.
 #' 
 #' @param iFile The input rdf file that will be read into R.
+#' @param rdf Boolean; if `TRUE`, then an rdf object is returned. If `FALSE`, 
+#'   then a character vector is returned.
 #' 
-#' @return An rdf object.
+#' @return An rdf object or character vector.
 #'   
 #' @examples
-#' zz <- read.rdf(system.file(
+#' zz <- read_rdf(system.file(
 #'   'extdata/Scenario/ISM1988_2014,2007Dems,IG,Most', 
 #'   "KeySlots.rdf", 
 #'   package = "RWDataPlyr"
@@ -158,7 +160,7 @@ read_rdf_run <- function(rdf.mat, rdf.obj)
 #' 
 #' @export
 
-read.rdf <- function(iFile)
+read.rdf <- function(iFile, rdf = TRUE)
 {
   stopifnot(tools::file_ext(iFile) == "rdf")
   stopifnot(file.exists(iFile))
@@ -171,6 +173,10 @@ read.rdf <- function(iFile)
     header = FALSE, 
     data.table = FALSE
   ))
+  
+  if (!rdf) {
+    return(rdf.mat)
+  }
   
   rdf.obj$position <- 1 # initialize where to read from
   rdf.obj <- read_rdf_meta(rdf.mat, rdf.obj)
@@ -198,7 +204,7 @@ read.rdf2 <- function(iFile)
 {
   .Deprecated("read.rdf")
   
-  read.rdf(iFile)
+  read.rdf(iFile, rdf = TRUE)
 }
 
 #' @rdname read.rdf