+++ NEWS.md
@@ -1,3 +1,14 @@
+# RWDataPlyr 0.6.0.9000
+
+*In development*
+
+* `rw_scen_aggregate()` and `rdf_aggregate()` gain `cpp` arguments. If `TRUE`, then `rdf_to_rwtbl2()` is used instead of `rdf_to_rwtbl()`
+* new function: `rdf_to_rwtbl2()`
+* `read_rdf()` and `read.rdf()` gained an `rdf` argument. If `TRUE`, it returns an `rdf` object, otherwise it returns a character vector.
+* in `rdf_to_rwtbl()`, `scenario` is coerced into a character
+* deprecating `rdf_to_rwtbl()` in favor of `rdf_to_rwtbl2()`
+
+
 # RWDataPlyr 0.6.0
 
 *Released April 10, 2018*